+++ tests/integration/storage_read_generic.py
@@ -3,7 +3,7 @@ from .mocks import add_infra, generate_mock_session
 import cloudwanderer
 
 
-class TestStorageRead(TestStorageConnectorReadMixin):
+class StorageReadTestMixin(TestStorageConnectorReadMixin):
     """Class which handles testing the various possible argument combinations."""
 
     def __init__(self, *args, **kwargs):
@@ -13,6 +13,7 @@ class TestStorageRead(TestStorageConnectorReadMixin):
 
     def setUp(self):
         self.connector = self.connector_class()
+        self.connector.init()
         self.wanderer = cloudwanderer.CloudWanderer(
             boto3_session=generate_mock_session(),
             storage_connectors=[self.connector]
@@ -25,7 +26,10 @@ class TestStorageRead(TestStorageConnectorReadMixin):
         self.not_expected_urns = []
 
     def test_no_args(self):
-        result = list(self.connector.read_resources())
+        try:
+            result = list(self.connector.read_resources())
+        except self.valid_exceptions:
+            return
 
         for account_id in ['111111111111', '222222222222']:
             self.expected_urns.extend([
@@ -39,7 +43,10 @@ class TestStorageRead(TestStorageConnectorReadMixin):
         self.assert_has_matching_aws_urns(result, self.expected_urns)
 
     def test_account_id(self):
-        result = list(self.connector.read_resources(account_id='111111111111'))
+        try:
+            result = list(self.connector.read_resources(account_id='111111111111'))
+        except self.valid_exceptions:
+            return
 
         for account_id in ['111111111111']:
             self.expected_urns.extend([
@@ -63,7 +70,10 @@ class TestStorageRead(TestStorageConnectorReadMixin):
         self.assert_does_not_have_matching_aws_urns(result, self.not_expected_urns)
 
     def test_region(self):
-        result = list(self.connector.read_resources(region_name='eu-west-2'))
+        try:
+            result = list(self.connector.read_resources(region='eu-west-2'))
+        except self.valid_exceptions:
+            return
 
         for account_id in ['111111111111', '222222222222']:
             self.expected_urns.extend([
@@ -80,7 +90,10 @@ class TestStorageRead(TestStorageConnectorReadMixin):
         self.assert_does_not_have_matching_aws_urns(result, self.not_expected_urns)
 
     def test_service(self):
-        result = self.connector.read_resources(service='ec2')
+        try:
+            result = list(self.connector.read_resources(service='ec2'))
+        except self.valid_exceptions:
+            return
 
         for account_id in ['111111111111', '222222222222']:
             self.expected_urns.extend([
@@ -97,7 +110,10 @@ class TestStorageRead(TestStorageConnectorReadMixin):
         self.assert_does_not_have_matching_aws_urns(result, self.not_expected_urns)
 
     def test_resource_type(self):
-        result = list(self.connector.read_resources(resource_type='bucket'))
+        try:
+            result = list(self.connector.read_resources(resource_type='bucket'))
+        except self.valid_exceptions:
+            return
 
         for account_id in ['111111111111', '222222222222']:
             self.expected_urns.extend([
@@ -114,13 +130,16 @@ class TestStorageRead(TestStorageConnectorReadMixin):
         self.assert_does_not_have_matching_aws_urns(result, self.not_expected_urns)
 
     def test_resource_urn(self):
-        result = list(self.connector.read_resources(urn=cloudwanderer.AwsUrn(
-            account_id='222222222222',
-            region='us-east-1',
-            service='iam',
-            resource_type='group',
-            resource_id='test-group'
-        )))
+        try:
+            result = list(self.connector.read_resources(urn=cloudwanderer.AwsUrn(
+                account_id='222222222222',
+                region='us-east-1',
+                service='iam',
+                resource_type='group',
+                resource_id='test-group'
+            )))
+        except self.valid_exceptions:
+            return
 
         for account_id in ['111111111111', '222222222222']:
             self.not_expected_urns.extend([